+++ src/main/java/org/ojalgo/optimisation/integer/NodeSolver.java
@@ -38,6 +38,7 @@ import org.ojalgo.optimisation.IntermediateSolver;
 import org.ojalgo.optimisation.ModelEntity;
 import org.ojalgo.optimisation.UpdatableSolver;
 import org.ojalgo.structure.Structure1D.IntIndex;
+import org.ojalgo.type.TypeUtils;
 import org.ojalgo.type.context.NumberContext;
 import org.ojalgo.type.keyvalue.EntryPair;
 
@@ -49,6 +50,7 @@ public final class NodeSolver extends IntermediateSolver {
     private static final boolean DEBUG = false;
     private static final NumberContext DYNANISM = NumberContext.of(8);
     private static final NumberContext LIMIT = PRECISION.withMode(RoundingMode.FLOOR);
+    private static final NumberContext PARAMETERS = NumberContext.of(12);
     private static final NumberContext SCALE = NumberContext.of(14);
 
     NodeSolver(final ExpressionsBasedModel model) {
@@ -64,8 +66,6 @@ public final class NodeSolver extends IntermediateSolver {
     }
 
     boolean generateCuts(final ModelStrategy strategy, final ExpressionsBasedModel target) {
-
-        // TODO Needs to be generalised to also handle cases with negative (full range) variables
 
         if (!this.isSolved()) {
             return false;
@@ -121,10 +121,11 @@ public final class NodeSolver extends IntermediateSolver {
                         int mj = entityMap.indexOf(j);
                         double aj = equation.doubleValue(j);
                         if (!SCALE.isZero(aj)) {
+                            BigDecimal AJ = TypeUtils.toBigDecimal(aj, SCALE);
                             if (entityMap.isNegated(j)) {
-                                cut.add(mj, -aj);
+                                cut.add(mj, AJ.negate());
                             } else {
-                                cut.add(mj, aj);
+                                cut.add(mj, AJ);
                             }
                         }
                     }
@@ -137,7 +138,7 @@ public final class NodeSolver extends IntermediateSolver {
 
                             ModelEntity<?> entity = pair.getKey();
                             ConstraintType type = pair.getValue();
-                            BigDecimal coefficient = BigDecimal.valueOf(aj);
+                            BigDecimal coefficient = TypeUtils.toBigDecimal(aj, SCALE);
                             BigDecimal adjusted = entity.adjust(coefficient);
 
                             if (ConstraintType.LOWER.equals(type)) {
@@ -165,7 +166,7 @@ public final class NodeSolver extends IntermediateSolver {
                     }
 
                     if (DEBUG) {
-                        BasicLogger.debug("Cut: {} := {}", cut, cut.getLinearEntrySet());
+                        BasicLogger.debug("Cut {}: {} < {}", name, cut.getLowerLimit(), cut.getLinearEntrySet());
                     }
 
                     BigDecimal cRHS = cut.getLowerLimit();
@@ -178,7 +179,7 @@ public final class NodeSolver extends IntermediateSolver {
                     if (cRHS.abs().compareTo(violation) > 0) {
                         target.removeExpression(name);
                         if (DEBUG) {
-                            BasicLogger.debug("\tViolation small! {}", cRHS);
+                            BasicLogger.debug(1, "Violation small! {}", cRHS);
                         }
                         continue;
                     }
@@ -209,32 +210,32 @@ public final class NodeSolver extends IntermediateSolver {
                     if (DEBUG) {
                         BigDecimal cRatio = MissingMath.divide(cLargest, cSmallest);
                         BigDecimal cEvaluated = cut.evaluate(result);
-                        BasicLogger.debug("\tLargest={}, Smallest={}, Ratio={}: {} < {}", cLargest, cSmallest, cRatio, cRHS, cEvaluated);
+                        BasicLogger.debug(1, "Largest={}, Smallest={}, Ratio={}: {} < {}", cLargest, cSmallest, cRatio, cRHS, cEvaluated);
                     }
 
                     if (DYNANISM.isSmall(cLargest, cSmallest)) {
                         target.removeExpression(name);
                         if (DEBUG) {
-                            BasicLogger.debug("\tDynanism large! {} >> {}", cLargest, cSmallest);
+                            BasicLogger.debug(1, "Dynanism large! {} >> {}", cLargest, cSmallest);
                         }
                     } else if (model.checkSimilarity(cut)) {
                         target.removeExpression(name);
                         if (DEBUG) {
-                            BasicLogger.debug("\tCut similar to current constraint!");
+                            BasicLogger.debug(1, "Cut similar to current constraint!");
                         }
                     } else {
+
                         // Accept this cut!
+                        cut.enforce(PARAMETERS);
                         cut.tighten();
                         if (DEBUG) {
-                            BasicLogger.debug("\t{}", cut);
+                            BasicLogger.debug(1, "{}", cut);
+                            BasicLogger.debug(1, "{} < {}", cut.getLowerLimit(), cut.getLinearEntrySet());
                         }
-                    }
-                    if (DEBUG) {
-                        BasicLogger.debug("\t{} < {}", cut.getLowerLimit(), cut.getLinearEntrySet());
-                    }
 
-                    if (target.options.logger_detailed && target.options.logger_appender != null) {
-                        target.options.logger_appender.println("{}: {} < {}", name, cut.getLowerLimit(), cut.getLinearEntrySet());
+                        if (target.options.logger_detailed && target.options.logger_appender != null) {
+                            target.options.logger_appender.println("{}: {} < {}", name, cut.getLowerLimit(), cut.getLinearEntrySet());
+                        }
                     }
                 }
             }